refactor: create 페이지 모달 피드백 반영
